PLAT-735-3 Fixed Some Input-Related Issues
