+++ platform-workflow-module/src/main/java/com/softicar/platform/workflow/module/workflow/item/message/WorkflowItemMessageDiv.java
@@ -6,10 +6,10 @@ import com.softicar.platform.common.date.DayTime;
 import com.softicar.platform.dom.elements.DomDiv;
 import com.softicar.platform.dom.elements.bar.DomActionBar;
 import com.softicar.platform.dom.elements.button.popup.DomPopupButton;
+import com.softicar.platform.dom.elements.checkbox.DomCheckbox;
 import com.softicar.platform.dom.refresh.bus.IDomRefreshBusEvent;
 import com.softicar.platform.dom.refresh.bus.IDomRefreshBusListener;
 import com.softicar.platform.emf.EmfImages;
-import com.softicar.platform.emf.attribute.field.bool.EmfBooleanInput;
 import com.softicar.platform.emf.data.table.EmfDataTableDivBuilder;
 import com.softicar.platform.emf.data.table.IEmfDataTableDiv;
 import com.softicar.platform.emf.form.popup.EmfFormPopup;
@@ -30,7 +30,7 @@ import java.util.stream.Collectors;
 public class WorkflowItemMessageDiv extends DomDiv implements IDomRefreshBusListener {
 
 	private final AGWorkflowItem item;
-	private final EmfBooleanInput showTransitionsCheckbox;
+	private final DomCheckbox showTransitionsCheckbox;
 	private IEmfDataTableDiv<WorkflowItemMessageRow> tableDiv;
 
 	public WorkflowItemMessageDiv(AGWorkflowItem item) {
@@ -47,7 +47,7 @@ public class WorkflowItemMessageDiv extends DomDiv implements IDomRefreshBusList
 					.setLabel(WorkflowI18n.ADD_NEW_MESSAGE)
 					.setEnabled(WorkflowPermissions.OPERATOR.test(item.getWorkflow().getModuleInstance(), CurrentBasicUser.get())));
 		this.showTransitionsCheckbox = actionBar//
-			.appendChild(new EmfBooleanInput(true).setLabel(WorkflowI18n.SHOW_TRANSITIONS));
+			.appendChild(new DomCheckbox(true).setLabel(WorkflowI18n.SHOW_TRANSITIONS));
 		showTransitionsCheckbox.addChangeCallback(this::refresh);
 		refresh(null);
 	}
